+++ spec/ransack/helpers/form_helper_spec.rb
@@ -310,12 +310,13 @@ module Ransack
       end
 
       context 'view has existing parameters' do
-        before do
-          @controller.view_context.params[:exist] = 'existing'
-        end
+
         describe '#sort_link should not remove existing params' do
-          subject { @controller.view_context
-            .sort_link(
+
+          before { @controller.view_context.params[:exist] = 'existing' }
+
+          subject {
+            @controller.view_context.sort_link(
               Person.search(
                 { sorts: ['name desc'] },
                 search_key: 'people_search'
@@ -324,7 +325,32 @@ module Ransack
               controller: 'people'
             )
           }
+
           it { should match /exist\=existing/ }
+        end
+
+        context 'using real ActionController Parameter object',
+          if: ::ActiveRecord::VERSION::MAJOR > 4 do
+
+          describe '#sort_link should include search params' do
+
+            subject { @controller.view_context.sort_link(Person.search, :name) }
+
+            let(:params) {
+              ActionController::Parameters
+              .new({ 'q' => { name_eq: 'TEST' }, controller: 'people' })
+            }
+
+            before { @controller.instance_variable_set(:@params, params) }
+
+            it {
+              should match(
+                /people\?q(%5B|\[)name_eq(%5D|\])=TEST&amp;q(%5B|\[)s(%5D|\])
+                =name\+asc/x,
+              )
+            }
+          end
+
         end
       end